+++ package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.7.0",
+    "@react-native-community/geolocation": "^2.0.2",
     "@testing-library/jest-dom": "^5.15.1",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
@@ -12,12 +13,13 @@
     "dotenv": "^10.0.0",
     "http-proxy-middleware": "^2.0.1",
     "react": "^17.0.2",
-    "react-bootstrap": "^2.0.3",
+    "react-daum-postcode": "^3.0.1",
     "react-dom": "^17.0.2",
     "react-google-login": "^5.2.2",
     "react-naver-maps": "0.0.13",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
+    "react-usestateref": "^1.0.8",
     "sync-request": "^6.1.0",
     "url-parse": "^1.5.3",
     "web-vitals": "^1.1.2"
