+++ public/index.html
@@ -37,6 +37,7 @@
     />
 
     <!-- Bootstrap -->
+
     <link
       type="text/css"
       rel="stylesheet"
