+++ src/01_Body/help_board/modify_board/index.css
@@ -1,0 +1,3 @@
+#layout {
+    margin-top: 20px;
+}
