+++ src/01_Body/help_board/write_board/index.css
@@ -1,0 +1,3 @@
+#layout {
+    margin-top: 20px;
+  }
