--- src/gvhelp/index.js
@@ -1,33 +1,0 @@
-import React from "react";
-import { RenderAfterNavermapsLoaded, NaverMap } from "react-naver-maps";
-
-function NaverMapAPI() {
-  return (
-    <NaverMap
-      mapDivId={"maps-getting-started-uncontrolled"} // default: react-naver-map
-      style={{
-        width: "100%", // 네이버지도 가로 길이
-        height: "85vh", // 네이버지도 세로 길이
-      }}
-      defaultCenter={{ lat: 37.554722, lng: 126.970833 }} // 지도 초기 위치
-      defaultZoom={13} // 지도 초기 확대 배율
-    />
-  );
-}
-
-function GVHelpPage() {
-  return (
-    <div className="container">
-      <RenderAfterNavermapsLoaded
-        //ncpClientId={process.env.REACT_APP_Naver_API_KEY} // 자신의 네이버 계정에서 발급받은 Client ID
-        ncpClientId = "qrrth0qept"
-        error={<p>Maps Load Error</p>}
-        loading={<p>Maps Loading...</p>}
-      >
-        <NaverMapAPI />
-      </RenderAfterNavermapsLoaded>
-    </div>
-  );
-}
-
-export default GVHelpPage;
