--- src/index/index.js
@@ -1,211 +1,0 @@
-import "./index.css";
-import { Carousel } from "antd";
-import { Row, Col, Card, Avatar } from "antd";
-import {
-  EditOutlined,
-  EllipsisOutlined,
-  SettingOutlined,
-} from "@ant-design/icons";
-
-function IndexPage() {
-  const contentStyle = {
-    height: "240px",
-    color: "#fff",
-    lineHeight: "160px",
-    textAlign: "center",
-    background: "#364d79",
-  };
-
-  const { Meta } = Card;
-
-  return (
-    <div>
-      {/* <!-- container --> */}
-      <div className="container">
-        <div id="carousel">
-          <Carousel autoplay>
-            <div>
-              <h3 style={contentStyle}>1</h3>
-            </div>
-            <div>
-              <h3 style={contentStyle}>2</h3>
-            </div>
-            <div>
-              <h3 style={contentStyle}>3</h3>
-            </div>
-            <div>
-              <h3 style={contentStyle}>4</h3>
-            </div>
-          </Carousel>
-        </div>
-        <div id="help-card">
-          <div id="help-card-title">도움이 필요해요!</div>
-          <div className="site-card-wrapper">
-            <Row gutter={16}>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-            </Row>
-            <div className="margin"></div>
-            <Row gutter={16}>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-              <Col span={8}>
-                <Card
-                  cover={
-                    <img
-                      alt="example"
-                      src="https://gw.alipayobjects.com/zos/rmsportal/JiqGstEfoWAOHiTxclqi.png"
-                    />
-                  }
-                  actions={[
-                    <SettingOutlined key="setting" />,
-                    <EditOutlined key="edit" />,
-                    <EllipsisOutlined key="ellipsis" />,
-                  ]}
-                >
-                  <Meta
-                    avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
-                    title="Card title"
-                    description="This is the description"
-                  />
-                </Card>
-              </Col>
-            </Row>
-            <div className="margin"></div>
-          </div>
-        </div>
-        <div id="banner">
-          {/* <!-- 우울증 극복 프로그램 연결 --> */}
-          <div id="footer-banner" className="section">
-            {/* <!-- container --> */}
-            <div className="container">
-              {/* <!-- row --> */}
-              <div className="row">
-                <div className="col-md-12">
-                  <div className="footer-banner-content">
-                    <h2 className="banner-txt1">
-                      요즘 축축 처지고 의욕이 없어..
-                    </h2>
-                    <p className="banner-txt2">우울증이 의심되나요?</p>
-                    <a
-                      className="primary-btn cta-btn"
-                      href="https://www.cyber1388.kr:447/"
-                    >
-                      사이버 상담센터 바로가기
-                    </a>
-                  </div>
-                </div>
-              </div>
-              {/* <!-- /row --> */}
-            </div>
-            {/* <!-- /container --> */}
-          </div>
-          {/* <!-- /우울증 극복 프로그램 연결 --> */}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-export default IndexPage;
