--- src/navigation/index.css
@@ -1,53 +1,0 @@
-/*=========================================================
-	03 -> Navigation
-===========================================================*/
-
-#navigation {
-  background: #f4f4f4;
-  /*border-top: 3px solid #012758;*/
-}
-
-a {
-  color: black;
-}
-
-/*----------------------------*\
-      Main nav
-  \*----------------------------*/
-
-.main-nav > li + li {
-  margin-left: 10px;
-  margin-right: 10px;
-}
-
-.main-nav > li > a {
-  padding: 20px 0px;
-}
-
-.main-nav > li > a:hover,
-.main-nav > li > a:focus,
-.main-nav > li > a:active {
-  color: #012758;
-  font-weight: bold;
-  background-color: transparent;
-}
-
-.main-nav > li > a:after {
-  content: "";
-  display: block;
-  width: 0%;
-  height: 2px;
-  background-color: #012758;
-  -webkit-transition: 0.2s all;
-  transition: 0.2s all;
-}
-
-.main-nav > li > a:hover:after,
-.main-nav > li > a:focus:after,
-.main-nav > li.active > a:after {
-  width: 100%;
-}
-
-.header-ctn li.nav-toggle {
-  display: none;
-}
